+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Ponder Changelog
 - Big refactor to simplify property bindings & add return internal refs.
 - API improvements (#93).
 - std::variant mapper.
+- std::optional mapper.
 - Tests: Catch2 updated to 2.13.1
 
 ## 3.1